Fix building error on Cygwin

+++ http_parser.c
@@ -412,7 +412,7 @@ int http_message_needs_eof(http_parser *parser);
 static enum state
 parse_url_char(enum state s, const char ch, int is_connect)
 {
-  assert(!isspace(ch));
+  assert(!isspace((int) ch));
 
   switch (s) {
     case s_req_spaces_before_url:
